Fix the <U\x+> keysyms in the en_US.UTF-8 Compose file
Based on src/KeysymStr.c and src/StrKeysym.c and comments in
bugs #11930¹ and #5129² it is clear that <U100XXXXX> is invalid;
those should be in the form U plus the hex of the UCS Code Point.

The 0x01000000 is ORed in by the code.

This update fixes all of those.

1] https://bugs.freedesktop.org/show_bug.cgi?id=11930
2] https://bugs.freedesktop.org/show_bug.cgi?id=5129