Fix problem with <dead_acute> <c> in pt_BR.UTF-8
The <dead_acute> <C> and <dead_acute> <c> lines in the pt_BR UTF-8
Compose file show "Ç" and "ç" (c with cedilla accent) (akin to the
ISO 8859 pt_BR Compose file) as the string but specify the keysym
and comment for Ć and ć (c with acute accent).

This commit normalizes those two lines to match the specified string.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=4671

Signed-off-by: James Cloos <cloos@jhcloos.com>