patches to avoid gcc warnings for libX11 (#3)
Author is Peter Breitenlohner <peb@mppmu.mpg.de>
Bug #17946, attachment #19441

This patch avoids the two gcc warnings
	../../../../libX11-1.1.5/modules/im/ximcp/imRm.c:413: warning: assignment discards qualifiers from pointer target type
	../../../../libX11-1.1.5/modules/im/ximcp/imRm.c:450: warning: assignment discards qualifiers from pointer target type

Note, that this as a rather crude fix of the problem (and it is really a
shame to cast name_table to non-const).

The right solution would be to declare XIMValuesList.supported_values
(in include/X11/Xlib.h) as 'const char **' (or '_Xconst char **').
This will, however, require extensive modifications in various places.