Revert "Change masculine to ordmasculine in Compose file comments"
This reverts commit 892b401d5acc055803a20e349ede0d64490f2230.

As Julien Cristau correctly points out, I misread the Compose file grammar
in modules/im/ximcp/imLcPrs.c, and those are keysyms, not comments, and
the keysym is named XK_masculine in keysymdef.h.   This change is thus a
bug in the Solaris compose tables to be fixed, not an improvement to bring
upstream.