Thai XIM not retrieve MB surrounding on UTF-8 LC
On th_TH.UTF-8 locale, Thai XIM rejects all combining characters for GTK+ apps
that use X Input Method.

This is because GTK+ imxim immodule passes surrounding text in locale encoding,
which is UTF-8 for UTF-8 locales. But current Thai XIM in Xlib assumes the
multi-byte StringConversionText response for the StringConversionCallback to
always be TIS-620, by retrieving a single byte and using it as-is.

If the Thai XIM tries to convert the multi-byte text based on locale codeset
before using it, it will work again.

X.Org But 12759 <http://bugs.freedesktop.org/show_bug.cgi?id=12759>

Signed-off-by: Theppitak Karoonboonyanan <thep@linux.thai.net>
Signed-off-by: Julien Cristau <jcristau@debian.org>