Add some (Serbian) Cyrillic NFD sequences.
A number of characters in use in the various countries which use the
Cyrillic script do not appear as pre-composed characters in The UCS
or Unicode; they are only available as combining-character sequences.

This commit adds support for using (prefix) dead keys and Multi_key-
initiated sequences to enter a number of these combining-character
sequences.  This ensures that users can enter these scripts even
when using the current Cyrillic keymaps, which lack support for
the combining characters.

Please see the discussions on the xkb mailing list.

Signed-off-by: James Cloos <cloos@jhcloos.com>