specs: move indexterm from glossdef to glossterm
This move fixes a Java class cast exception in the glossary.
The problem was introduced in commit
26f4f0d50840fe5ba4c46aae0a8e68db0059434b

It may not happen on all versions of the doc toolchain.
There is no reason why indexterm cannot appear in glossdef,
this is a workaround to an implementation problem found by
trial and error.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>
Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>