nls: Switch one of the interrobang sequences to gnaborretni
Since gnaborretni is primarily used in LTR locales, the ? part of the
sequence reflects the position of the bulge in the ?.  When scanning an
inverted interrobang left-to-right, you see the bulge first, so the
compose sequence ?! is used; upright interrobang shows the bulge
last, so it's !?.

This is a change in behaviour, but I was unable to come up with anything
better.  ??!! or !!?? would match the patterns for ¿ and ¡, but they
would delay evaluation of them; "¿Qué?" wouldn't show the ¿ until you hit
Q.  Likewise ?!?! and !?!? would delay showing the interrobang itself.
~!? and ~?! were considered but are arguably less intuitive.

Reviewed-by: James Cloos <cloos@jhcloos.com>
Signed-off-by: Adam Jackson <ajax@redhat.com>
(cherry picked from commit f70527f05677aaf6a493ba42d2d95f738aaebafa)