Fix use-after-free in _XReply on X errors.
_XReply would always call dequeue_pending_request on errors.  When it
got an error for the current request, it would call dequeue, then break
out of the loop; then, if it had an error in the event queue, it would
compare it with the sequence number of the now-freed pending request.
_XReply already stored that sequence number in dpy->last_request_read
before freeing it, so look at that instead.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=29412

Signed-off-by: Jamey Sharp <jamey@minilop.net>
Signed-off-by: Josh Triplett <josh@joshtriplett.org>
(cherry picked from commit 4b8ff7db39f2fe7ef12968d462aaf3f9054b6c18)