config: remove man-pages configuration option
This option was added in commit 6e752ea120 with no explanation.
The section number is provoded by XORG_MANPAGE_SECTIONS
There is no case where libX11 should be different than other libs
The option was also used to disable building of the man pages,
which build in 14 secs. No indication this is required.

If there is a requirement from system builders to disable building
of man pages, it could be done consistently for all modules.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>