docs: Remove directory prerequisites from make rules
Make expects prerequisites to be files with valid timestamps, and
directories are treated as always being out of date. Thus, any targets
depending on directories will always be rebuilt.

Instead, the doc rules are changed to always create the target's leading
directory. This should prevent the documentation from being rebuilt when
"make install" is run.

Signed-off-by: Dan Nicholson <dbn.lists@gmail.com>
Reviewed-by: Jeremy Huddleston <jeremyhu@apple.com>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>