compose specs: generate chunked html
For large 600 page documents such as this one, chunked html
provide faster browser load time and better navigation.

Simply click on the locale of your choice in the toc and
the browser loads just that one file. Being a DocBook,
it benefits from all of the usual features and can be
integrated with the rest of the documentation.

Acked-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>