libX11 specs: move </para> above <varaiablelist>
Many, but not all, function synopsis have a paragraph that
nests the variable list. The code was generated by doclifter
and there is no apprent reasons as to why it was done this way.
Found while investigating a spacing issue.

Before the patch:
<para>Some blurb about the function
<funcsynopsis>
</funcsynopsis>
<variablelist>
</variablelist>
</para>

After the patch:
<para>Some blurb about the function</para>
<funcsynopsis>
</funcsynopsis>
<variablelist>
</variablelist>

There are no noticable differences when reading the doc
other than than the removed few pixels of spacing between
the function sysnopsis and the variable list block.

In some cases, there are no "blurb about the fucntion"
and the empty paragraph is removed.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>