docbook.am: do not generate docs if docbook customization layer is missing
The stylesheets used to be only about style, fonts or colors.
Complex features are now used like olink and chunked html which
may cause a build break when stylesheets are missing or lead to hard
to find problem.

Some modules may be built when stylesheets are present while others
are built without. There is no requirement to build crippled docs.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>