docbook.am: add search path for local entities
Currently, only $(XORG_SGML_PATH)/X11 is searched for xml entities.
A module may want to add entities that apply only to itself and
not to all modules, like the xserver does.

This feature may or may not be used in this module, but all modules
share a copy of docbook.am.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>