compose: upgrade makefile to support olinking on chunked html
The essential differences over the regular docbook.am are:
Adding root.filename parameter for naming of chapters html files.
Using xhtml xmlto format and xorg-chunk.xsl stylesheet
Set olink.base.uri for pdf but not for chunked html
Olink is not applicable to ps and txt formats.

Html chapters are added to shelf_DATA as they are also installed.
The xml is generated from a perl script and not distributed.

Requires version 1.10 of xorg-sgml-doctools.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>