specs: handle multiple sets of copyright notice/license/warranty
Docbook groups all the <copyright> elements together and all the
<legalnotice> elements together.

We cannot have a sequence:
<copyright> <legalnotice> <copyright> <legalnotice> [...]

A workaround, which was done in some documents, is to put the copyright
notice inside the legalnotice in plain text without the <copyright> element.
A formal paragraph title is added here which makes the copyright notice bold,
and makes it much easier to locate.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>