Block for other threads in _XUserLockDisplay
Wait for all other threads to release the user-level lock when
acquiring it. This ensures that only one thread at a time holds the
user-level lock, necessary as it is a nesting lock and a single
variable is used to determine when the lock is nesting and when it is
contended.

Signed-off-by: Keith Packard <keithp@keithp.com>
Reviewed-by: Jamey Sharp <jamey@minilop.net>