Fix the gtk+ additions
(Some of) the Dstroke and dstroke entries already were present as U011[01],
even though XK_Dstroke and XK_dstroke are part of the latin2 set in keysymdef.h.

The addition of <Multi_key> <o> <apostrophe> as a postfix version of
<Multi_key> <apostrophe> <o> blocks the existing entries for ǻ and Ǻ.
That prevents its and <Multi_key> <O> <apostrophe>’s addition.

Signed-off-by: James Cloos <cloos@jhcloos.com>