Allow overriding location of keysymdef.h
Currently keysymdef.h is found by using the includedir of xproto.  This doesn't
work when cross-compiling with a sysroot as that ends up being /usr/include/X11,
not a path into the cross-build environment.

So, add an option to allow explicitly specifying the location of keysymdef.h,
and verify that the specified or found path exists.

(original patch by Martin Jansa <martin.jansa@gmail.com>, revised by myself)

Signed-off-by: Ross Burton <ross.burton@intel.com>
Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>
Reviewed-by: Daniel Stone <daniel@fooishbar.org>